+++ src/main/java/com/phu/backend/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.phu.backend.config.redis;
+package com.phu.backend.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
