+++ src/main/java/com/phu/backend/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.phu.backend.config.swagger;
+package com.phu.backend.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Info;
