+++ src/main/java/com/phu/backend/security/filter/jwt/JWTFilter.java
@@ -1,5 +1,6 @@
-package com.phu.backend.config.jwt;
+package com.phu.backend.security.filter.jwt;
 
+import com.phu.backend.security.util.jwt.JWTUtil;
 import com.phu.backend.dto.auth.MemberDetails;
 import com.phu.backend.exception.jwt.AccessTokenExpiredException;
 import com.phu.backend.exception.jwt.TokenNotValidateException;
