+++ src/main/java/com/phu/backend/security/oauth/GoogleResponse.java
@@ -1,4 +1,4 @@
-package com.phu.backend.config.oauth;
+package com.phu.backend.security.oauth;
 
 import lombok.RequiredArgsConstructor;
