+++ src/main/resources/application-local.yml
@@ -31,12 +31,13 @@ spring:
           google:
             client-secret: ${CLIENT_SECRET}
             client-id: ${CLIENT_ID}
+            redirect-uri: http://localhost:8080/login/oauth2/code/google
             scope:
               - email
               - profile
 
 jwt:
   secret: asdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdf12
-  access-token-expiration-minutes: 1
+  access-token-expiration-minutes: 600000
   refresh-token-expiration-minutes: 6000000
